Update index.js resolved Mongodb connection error
Resolved MongooseServerSelectionError: Could not connect to any servers in your MongoDB Atlas cluster indicates that your server is unable to establish a connection to your MongoDB Atlas cluster. As the error message suggests, this is often due to your IP address not being whitelisted in your MongoDB Atlas security settings.